✨ Enables passwordless user creation
Implements functionality to create users without passwords for development and testing purposes.
Adds a new API endpoint for developers to automatically send emails to users of a specific role who do not have passwords.
Also includes a link to open account in email
Fixes a bug in mail processor that prevented sending mails